+++ src/main/java/com/ayakacraft/carpetayakaaddition/CarpetAyakaAddition.java
@@ -21,7 +21,7 @@
 package com.ayakacraft.carpetayakaaddition;
 
 import carpet.CarpetServer;
-import com.ayakacraft.carpetayakaaddition.utils.mods.ModUtils;
+import com.ayakacraft.carpetayakaaddition.utils.ModUtils;
 import com.ayakacraft.carpetayakaaddition.utils.translation.AyakaLanguage;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;