+++ src/main/java/com/ayakacraft/carpetayakaaddition/mixin/rules/betterMobCap/PatrolSpawnerMixin.java
@@ -1,0 +1,50 @@
+/*
+ * This file is part of the Carpet Ayaka Addition project, licensed under the
+ * GNU General Public License v3.0
+ *
+ * Copyright (C) 2025  Calboot and contributors
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package com.ayakacraft.carpetayakaaddition.mixin.rules.betterMobCap;
+
+import com.ayakacraft.carpetayakaaddition.helpers.rules.BetterMobCapHelper;
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
+import net.minecraft.entity.EntityType;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.server.network.ServerPlayerEntity;
+import net.minecraft.world.spawner.PatrolSpawner;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+
+@Mixin(PatrolSpawner.class)
+public class PatrolSpawnerMixin {
+
+    @WrapOperation(
+            method = "spawn",
+            at = @At(
+                    value = "INVOKE",
+                    target = "Lnet/minecraft/entity/player/PlayerEntity;isSpectator()Z"
+            )
+    )
+    private boolean applyMobCaps(PlayerEntity instance, Operation<Boolean> original) {
+        if (BetterMobCapHelper.checkBelowCaps((ServerPlayerEntity) instance, EntityType.PILLAGER)) {
+            return original.call(instance);
+        }
+        return true;
+    }
+
+}