+++ src/main/java/com/ayakacraft/carpetayakaaddition/mixin/rules/fakePlayerResidentBackupFix/MinecraftServerMixin.java
@@ -21,8 +21,8 @@
 package com.ayakacraft.carpetayakaaddition.mixin.rules.fakePlayerResidentBackupFix;
 
 import com.ayakacraft.carpetayakaaddition.CarpetAyakaSettings;
-import com.ayakacraft.carpetayakaaddition.utils.mods.GcaHelper;
-import com.ayakacraft.carpetayakaaddition.utils.mods.ModUtils;
+import com.ayakacraft.carpetayakaaddition.helpers.mods.GcaHelper;
+import com.ayakacraft.carpetayakaaddition.utils.ModUtils;
 import me.fallenbreath.conditionalmixin.api.annotation.Condition;
 import me.fallenbreath.conditionalmixin.api.annotation.Restriction;
 import net.minecraft.server.MinecraftServer;