+++ changelog.md
@@ -1,3 +1,4 @@
 ### General
 
+- ***Added*** carpet rule `betterMobCap` to limit phantom and pillager patrol spawning
 - Updated translations